Toolbar stuff
Made the toolbar more coherant, and it is now owned by the application
view. It also has a selector for classrooms now, as well as 3 buttons:
list, new and save.

List will filter which placements you'll see in the list view, and will
someday (not yet) take you back to the list view from the placement
workbench.

New will someday (not yet) create a new placement for the selected
classroom, and take you to its workbench.

Save will someday (not yet) do what it used to do (save a placement).
Right now it's just broken.

To populate the selector, I had Rails generate a big dumb javascript
object in a script tag in index.html. To do so, I had to convert index
to be erb, which means that underscore is now using # instead of % for
interpolation et all.