Get undo-redo working
The problem we were hitting before was that the Placement loaded data
from Rails asynchronously, but the PlacementView was created
immediately. This had two effects: only the unplaced-stuents company
existed to track when the undomanager was set up, and then when the data
finally arrived all of the adds and removes to unplaced students ended
up being tracked (even though that was the "initial" placement).

Solution is to build the PlacementView in the success callback for
Placement.fetch.

One weird thing: we need to call undo twice! Seems like this is because
the undomanager is picking up the 'select' event on the Students, even
though I didn't explicitly tell it to track events on students. Oh well.

I also added keyboard shortcuts: cmd or ctrl + z for undo, and that plus
shift for redo.